requirements.txt
python-dotenv
langchain
langchain_openai
langchain_huggingface
langchain_google_genai